ARM: mm: fix support for HW coherent systems in PL310 cache
When a PL310 cache is used in a system that provides hardware coherency,
the entire outer cache operations are useless, and can be skipped.
Moreover, on some systems, it is harmful as it causes deadlocks between
the Marvell coherency mechanism, the Marvell PCIe controller and the
Cortex-A9.

This commit extends a previous commit:
'commit 98ea2dba6593 ("ARM: 8076/1: mm: add support for HW coherent systems
in PL310 cache")' which added the io-coherent support for the PL310 cache
by also disabling the outer cache flush range operation.

In the current kernel implementation, the outer cache flush range
operation is triggered by the dma_alloc function.
This operation can be take place during runtime and in some circumstances
may lead to the PCIe/PL310 deadlock on Armada 375/38x SoCs.

Signed-off-by: Nadav Haklai <nadavh@marvell.com>
Change-Id: I253ac4dcb323627edf59cbabad42c01650b1ce7f
Reviewed-on: http://vgitil04.il.marvell.com:8080/19932
Reviewed-by: Ofer Heifetz <oferh@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/28160