ARM: mvebu: Enable SCU Speculative linefills to L2
When set, coherent linefill requests are sent speculatively to the L2C-310
in parallel with the tag look-up. If the tag look-up misses, the confirmed
linefill is sent to the L2C-310 and gets RDATA earlier because the data
request was already initiated by the speculative request.

This feature may improve overall system performance.

Change-Id: Ib7b29668bace0f09762a7cbab3ea149bc776891c
Signed-off-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/23491
Reviewed-by: Neta Zur Hershkovits <neta@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/23672
Reviewed-on: http://vgitil04.il.marvell.com:8080/27307
Tested-by: Lior Amsalem <alior@marvell.com>