ARM: mvebu: Disable CPU freq on Armada 38x for the multi CPU flavors
On multi CPU Armada 38x SoCs, under heavy I/O load, the system hangs when
cpufreq is enabled.
Waiting for a solution to this issue, this patch disables the cpufreq
support for those SoC.

The single CPU flavor of Armada 38x, Armada 380 does not suffer from the
cpufreq related system hang and therefore cpufreq can be enabled on this
platform.

Change-Id: Iae35574c2d538c01e22206724cf708691ba717b7
Signed-off-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/23688
Tested-by: Star_Automation <star@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/27314
Tested-by: Lior Amsalem <alior@marvell.com>