axim: fix nr_chan caculcation for AXIM v1
AXI monitor v1 doesn't have the version register so
we can't get the channel count and other info.
When reading from the non-existent register, the v1 monitor
return 0 so we use this behavior to emulate a register read.
the only value that needs to be fixed is the channel count
that equals to 4 for v1

Change-Id: I8c0f74616d7be67592be3bdd5ac2db6817628555
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/31046
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Tested-by: Star_Automation <star@marvell.com>