pinctrl: cp110: add pinctrl support for CP0 and CP1 which exist in Armada 8K
Differences between Armada-80x0 (dual CP) and Armada-70x0 (single CP):
- CP present in A7K has access to control all available MPP's 0-62
- CP0 present in A8K has access to control MPP's 32-62
- CP1 present in A8K has access to control MPP's 0-32

Despite mentioned differences the base address of CP0 (Armada-80x0) and
CP (Armada-70x0) is the same. Therefore CP and CP1 share the same
*ctrl_get and *ctrl_set functions and differences between them are handled
with "variant" flag. The CP1 has it's separate base address and
*ctrl_get and *ctrl_set functions.

Additional notes:
The cp110_1_mpp_controls has to have defined 62 MPP because it is used
together with 'armada_cp110_mpp_modes' table, which has 62 entries.
Nevertheless as mentioned earlier, thanks to the 'variant' flag only
MPP 0-31 are valid for CP1. Accordingly, the same is for the CP0:
62 MPP are defined but for Armada 8K only 32-62 are valid while for
Armada 7K all 62 are valid.

The armada_cp110_0_mpp_gpio_ranges is used for both Armada 8K and Armada
7K, and again "variant" flag is used for validating if appropriate MPP
is available for each platform. It will be more clear to define separate
pinctrl_gpio_range for Armada 7K (0-62) and Armada 8K (32-62), but the
common gpio framework expects that the gpio starts from 0.

Change-Id: I3ca60d24c2ef318e9a277931858539d90670f9c5
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32374
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>