mmc: sdhci-xenon: support tuning step divider property in DTS
This DTS property is used to set tuning step divider.
It will be default to 64 if do not set it in DTS.
HW DLL will generate the tuning step value, then the driver
applies it to tuning step register after divider.
The divider value is based on testing.
Now all the SoC using XENON IP work at common divider value 64.
However since the timing is also relevant to board, it is better to
provide this configuration in DTS for customer to test and find the right
divider value if the common divider value 64 does not work on their
platform.

Change-Id: I3a5cde545b47ea7d0de25f8d0726dc5eb83d9059
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32168
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>