fix: spi: a3700: support 4 byte in each serial IO transfer for fifo mode
In order to increase SPI read/write performance, the 4 byte length
feature should be enabled to read or write 4 byte in each serial IO
transfer.
This patch sets the byte length to 1 byte in none fifo mode, and
set to 4 byte mode in fifo mode.
The reading and writing of none 4 byte aligned length is also supported
by it.
According to the testing, the 4 byte length transfer combined with higher
SPI clock frequency such as 100MHZ could increase the SPI performance.

Change-Id: I90b731f3cdf52e05061233d040c10380815246cc
Signed-off-by: zachary <zhangzg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/33634
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>