fix: xor_v2: disable interrupt coalescing feature
This patch disables the coalescing of XOR driver - because if the driver
enables this feature without a timer threshold, it would cause that some of
the descriptors will not get an interrupt and cause kernel panic.
The message threshold in XOR driver is set to 0 by default, that mean on
every descriptor will cause an interrupt.

Change-Id: Id0c2d15225c634ae0251487fa9e4789de5a27640
Signed-off-by: Hanna Hawa <hannah@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/34066
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>