fix: arm64: exception: ignore EL0 asynchronous external abort
There is an EL0 asynchronous external abort on some boot cases,
during file-system mount stage. The root cause of the abort is not
clear for now.

This patch adds a WA which allows to ignore EL0 asynchronous
external abort (in addition to EL1 asynchronous external abort
which is already ignored).

Change-Id: I317ddfc415b8a8c4ed7b087a50087fc2e7523b88
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/34345
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>