fix: rtc: a80x0: fix rtc drift
This patch fixes the RTC drift between System Clock ('date')
and RTC Clock ('hwclock') as the RTC clock was too slow.

The RTC requires external oscillator.
The problem was that, in 80x0, CP0 RTC clock was used while CP1 RTC was
the one which is connected (by package) to the oscillator.
CP0 RTC is not connected at all (though the IP can be accessed).

This patch enables CP0 RTC for 70x0 and CP1 RTC for 80x0.

(Note: armada-cp110-1.dtsi was not updated as the RTC node there
is already disabled).

Change-Id: Ieeb5d0256821ddbcb51c149fde8429388d30e053
Signed-off-by: Haim Boot <hayim@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/34581
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>